Created NON-CRUD Test, Other test, Created WEB-API Created Console menu

+++ H8GXCF_HFT_2022231.Logic/Services/MemberLogic.cs
@@ -9,26 +9,40 @@ using System.Threading.Tasks;
 
 namespace H8GXCF_HFT_2022231.Logic.Services
 {
-    public class MemberLogic : IMemberLogic
+    public class MemberLogic : ILogic<Member>
     {
         IMemberRepository memberRepository;
         public MemberLogic(IMemberRepository memberRepository)
         {
             this.memberRepository = memberRepository;
         }
-        public void Create(Member member)
+        public void Create(Member item)
         {
-            memberRepository.Create(member);
+            if (item.Name.Length < 3)
+            {
+                throw new ArgumentException("Member name was too short...");    
+            }
+            memberRepository.Create(item);
         }
 
         public void Delete(int id)
         {
+            var member = memberRepository.Read(id);
+            if (member == null)
+            {
+                throw new ArgumentException("Member does not exists...");
+            }
             memberRepository.Delete(id);
         }
 
         public Member Read(int id)
         {
-           return memberRepository.Read(id);
+            var member = memberRepository.Read(id);
+            if (member == null)
+            {
+                throw new ArgumentException("Member does not exists...");
+            }
+            return memberRepository.Read(id);
         }
 
         public IEnumerable<Member> ReadAll()
@@ -36,19 +50,35 @@ namespace H8GXCF_HFT_2022231.Logic.Services
             return memberRepository.ReadAll();
         }
 
-        public void Update(Member member)
+        public void Update(Member item)
         {
-           memberRepository.Update(member);
+            if (item == null)
+            {
+                throw new ArgumentException("member does not exists...");
+            }
+           memberRepository.Update(item);
         }
-        public double AVGAge()
+        public Dictionary<string, int> MaleFemaleCount()
         {
-            return memberRepository.ReadAll().Average(t => t.Age);
+            var result = from x in memberRepository.ReadAll()
+                         group x by x.Gender.ToString() into g
+                         select new
+                         {
+                             g.Key,
+                             Count = g.Count()
+                         };
+            return result.ToDictionary(x => x.Key, x => x.Count);
         }
-        public List<string> ActiveMembers()
+        public Dictionary<string,int> InstructorClientCount ()
         {
-            return memberRepository.ReadAll()
-                .Where(t => t.Membership.Active)
-                .Select(t => t.Name).ToList();
+            var result = from x in memberRepository.ReadAll()
+                         group x by x.Instructor.Name into g
+                         select new
+                         {
+                             g.Key,
+                             Count = g.Count()
+                         };
+            return result.ToDictionary(x => x.Key, x => x.Count);
         }
         public Dictionary<string,int> MemberTypeCount()
         {
@@ -60,6 +90,36 @@ namespace H8GXCF_HFT_2022231.Logic.Services
                        Count = g.Count()
                    };
             return result.ToDictionary(x => x.Key, x => x.Count);
+        }
+        public Dictionary<string, double> AverageFeeByGender()
+        {
+            var result = from x in memberRepository.ReadAll()
+                         group x by x.Gender.ToString() into g
+                         select new
+                         {
+                             g.Key,
+                             Avarage = g.Average(x => x.Membership.SignupFee)
+                         };
+            return result.ToDictionary(x => x.Key, x => x.Avarage);
+        }
+        public Dictionary<int, double> ActiveMembersAverageAgeAndCount()
+        {
+            var result = from x in memberRepository.ReadAll()
+                         group x by x.Membership.Active into g where g.Key == true
+                         select new
+                         {
+                             Count = g.Count(),
+                             Average = g.Average(x => x.Age)
+                         };
+            return result.ToDictionary(x => x.Count, x => x.Average);
+        }
+        public double AverageMemberAge()
+        {
+            return memberRepository.ReadAll().Average(t => t.Age);
+        }
+        public int MemberCount()
+        {
+            return memberRepository.ReadAll().Count();
         }
     }
 }
